arba_cms/requirements.txt
asgiref==3.8.1
boto3==1.36.9
botocore==1.36.9
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
jmespath==1.0.1
Markdown==3.7
mysqlclient==2.2.7
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
